refactor submitCheque
