+++ truffle-config.js
@@ -68,7 +68,7 @@ module.exports = {
   // Configure your compilers
   compilers: {
     solc: {
-      version: "0.5.8",    // Fetch exact version from solc-bin (default: truffle's version)      
+      version: "0.5.10",    // Fetch exact version from solc-bin (default: truffle's version)      
       settings: {          // See the solidity docs for advice about optimization and evmVersion
        optimizer: {
          enabled: true,
